requirements.txt
requests>=2.31.0
websocket-client>=1.6.0
langdetect>=1.0.9
python-dotenv>=1.0.0 